WCM-694: move retract cronjob to cli, to keep consistent with all the other cli commands
